Add new mobile entrypoint for rn
