+++ shared.ts
@@ -1,0 +1,402 @@
+import { load, type Cheerio, type CheerioAPI, type Element } from "cheerio";
+import { CONSTANTS } from "./constants";
+
+export interface ILinkPreviewResponse {
+  url: string;
+  title: string;
+  siteName: string | undefined;
+  author: string | undefined;
+  description: string | undefined;
+  mediaType: string;
+  contentType: string | undefined;
+  images: string[];
+  videos: IVideoType[];
+  favicons: string[];
+}
+
+export interface IVideoType {
+  url: string | undefined;
+  secureUrl: string | null | undefined;
+  type: string | null | undefined;
+  width: string | undefined;
+  height: string | undefined;
+}
+
+/**
+ * Options common to every entrypoint (Node, mobile/browser, ...). Platform-specific
+ * options (like Node's `resolveDNSHost`) are added on top of this by each entrypoint,
+ * since they don't make sense - or aren't implementable - everywhere.
+ */
+export interface ILinkPreviewBaseOptions {
+  headers?: Record<string, string>;
+  imagesPropertyType?: string;
+  proxyUrl?: string;
+  timeout?: number;
+  followRedirects?: `follow` | `error` | `manual`;
+  handleRedirects?: (baseURL: string, forwardedURL: string) => boolean;
+  onResponse?: (response: ILinkPreviewResponse, doc: CheerioAPI, url?: URL) => ILinkPreviewResponse;
+}
+
+export interface IPreFetchedResource {
+  headers: Record<string, string>;
+  status?: number;
+  imagesPropertyType?: string;
+  proxyUrl?: string;
+  url: string;
+  data: string;
+}
+
+function metaTag(doc: CheerioAPI, type: string, attr: string) {
+  const nodes = doc(`meta[${attr}='${type}']`);
+  return nodes.length ? nodes : null;
+}
+
+function metaTagContent(doc: CheerioAPI, type: string, attr: string) {
+  return doc(`meta[${attr}='${type}']`).attr(`content`);
+}
+
+function getTitle(doc: CheerioAPI) {
+  let title =
+    metaTagContent(doc, `og:title`, `property`) || metaTagContent(doc, `og:title`, `name`);
+  if (!title) {
+    title = doc(`head > title`).text();
+  }
+  return title;
+}
+
+function getSiteName(doc: CheerioAPI) {
+  const siteName =
+    metaTagContent(doc, `og:site_name`, `property`) || metaTagContent(doc, `og:site_name`, `name`);
+  return siteName;
+}
+
+function getAuthor(doc: CheerioAPI) {
+  const author =
+    metaTagContent(doc, `author`, `name`) || metaTagContent(doc, `article:author`, `property`);
+  return author;
+}
+
+function getDescription(doc: CheerioAPI) {
+  const description =
+    metaTagContent(doc, `description`, `name`) ||
+    metaTagContent(doc, `Description`, `name`) ||
+    metaTagContent(doc, `og:description`, `property`);
+  return description;
+}
+
+function getMediaType(doc: CheerioAPI) {
+  const node = metaTag(doc, `medium`, `name`);
+  if (node) {
+    const content = node.attr(`content`);
+    return content === `image` ? `photo` : content;
+  }
+  return metaTagContent(doc, `og:type`, `property`) || metaTagContent(doc, `og:type`, `name`);
+}
+
+function getImages(doc: CheerioAPI, rootUrl: string, imagesPropertyType?: string) {
+  let images: string[] = [];
+  let nodes: Cheerio<Element> | null;
+  let src: string | undefined;
+  let dic: Record<string, boolean> = {};
+
+  const imagePropertyType = imagesPropertyType ?? `og`;
+  nodes =
+    metaTag(doc, `${imagePropertyType}:image`, `property`) ||
+    metaTag(doc, `${imagePropertyType}:image`, `name`);
+
+  if (nodes) {
+    nodes.each((_: number, node: Element) => {
+      if (node.type === `tag`) {
+        src = node.attribs.content;
+        if (src) {
+          src = new URL(src, rootUrl).href;
+          images.push(src);
+        }
+      }
+    });
+  }
+
+  if (images.length <= 0 && !imagesPropertyType) {
+    src = doc(`link[rel=image_src]`).attr(`href`);
+    if (src) {
+      src = new URL(src, rootUrl).href;
+      images = [src];
+    } else {
+      nodes = doc(`img`);
+
+      if (nodes?.length) {
+        dic = {};
+        images = [];
+        nodes.each((_: number, node: Element) => {
+          if (node.type === `tag`) src = node.attribs.src;
+          if (src && !dic[src]) {
+            dic[src] = true;
+            // width = node.attribs.width;
+            // height = node.attribs.height;
+            images.push(new URL(src, rootUrl).href);
+          }
+        });
+      }
+    }
+  }
+
+  return images;
+}
+
+function getVideos(doc: CheerioAPI) {
+  const videos = [];
+  let nodeTypes;
+  let nodeSecureUrls;
+  let nodeType;
+  let nodeSecureUrl;
+  let video;
+  let videoType;
+  let videoSecureUrl;
+  let width;
+  let height;
+  let videoObj;
+  let index;
+
+  const nodes = metaTag(doc, `og:video`, `property`) || metaTag(doc, `og:video`, `name`);
+
+  if (nodes?.length) {
+    nodeTypes = metaTag(doc, `og:video:type`, `property`) || metaTag(doc, `og:video:type`, `name`);
+    nodeSecureUrls =
+      metaTag(doc, `og:video:secure_url`, `property`) ||
+      metaTag(doc, `og:video:secure_url`, `name`);
+    width =
+      metaTagContent(doc, `og:video:width`, `property`) ||
+      metaTagContent(doc, `og:video:width`, `name`);
+    height =
+      metaTagContent(doc, `og:video:height`, `property`) ||
+      metaTagContent(doc, `og:video:height`, `name`);
+
+    for (index = 0; index < nodes.length; index += 1) {
+      const node = nodes[index];
+      if (node.type === `tag`) video = node.attribs.content;
+
+      nodeType = nodeTypes?.[index];
+      if (nodeType?.type === `tag`) {
+        videoType = nodeType ? nodeType.attribs.content : null;
+      }
+
+      nodeSecureUrl = nodeSecureUrls?.[index];
+      if (nodeSecureUrl?.type === `tag`) {
+        videoSecureUrl = nodeSecureUrl ? nodeSecureUrl.attribs.content : null;
+      }
+
+      videoObj = {
+        url: video,
+        secureUrl: videoSecureUrl,
+        type: videoType,
+        width,
+        height,
+      };
+      if (videoType && videoType.indexOf(`video/`) === 0) {
+        videos.splice(0, 0, videoObj);
+      } else {
+        videos.push(videoObj);
+      }
+    }
+  }
+
+  return videos;
+}
+
+// returns default favicon (//hostname/favicon.ico) for a url
+function getDefaultFavicon(rootUrl: string): string {
+  return new URL(`/favicon.ico`, rootUrl).href;
+}
+
+// returns an array of URLs to favicon images
+function getFavicons(doc: CheerioAPI, rootUrl: string): string[] {
+  const images = [];
+  let nodes: Cheerio<Element>;
+  let src: string | undefined;
+
+  const relSelectors = [`rel=icon`, `rel="shortcut icon"`, `rel=apple-touch-icon`];
+
+  relSelectors.forEach((relSelector) => {
+    // look for all icon tags
+    nodes = doc(`link[${relSelector}]`);
+
+    // collect all images from icon tags
+    if (nodes.length) {
+      nodes.each((_: number, node: Element) => {
+        if (node.type === `tag`) src = node.attribs.href;
+        if (src) {
+          src = new URL(src, rootUrl).href;
+          images.push(src);
+        }
+      });
+    }
+  });
+
+  // if no icon images, use default favicon location
+  if (images.length <= 0) {
+    images.push(getDefaultFavicon(rootUrl));
+  }
+
+  return images;
+}
+
+function parseImageResponse(url: string, contentType: string) {
+  return {
+    url,
+    mediaType: `image`,
+    contentType,
+    favicons: [getDefaultFavicon(url)],
+  };
+}
+
+function parseAudioResponse(url: string, contentType: string) {
+  return {
+    url,
+    mediaType: `audio`,
+    contentType,
+    favicons: [getDefaultFavicon(url)],
+  };
+}
+
+function parseVideoResponse(url: string, contentType: string) {
+  return {
+    url,
+    mediaType: `video`,
+    contentType,
+    favicons: [getDefaultFavicon(url)],
+  };
+}
+
+function parseApplicationResponse(url: string, contentType: string) {
+  return {
+    url,
+    mediaType: `application`,
+    contentType,
+    favicons: [getDefaultFavicon(url)],
+  };
+}
+
+function parseTextResponse(
+  body: string,
+  url: string,
+  options: ILinkPreviewBaseOptions = {},
+  contentType?: string,
+): ILinkPreviewResponse {
+  const doc = load(body);
+
+  let response = {
+    url,
+    title: getTitle(doc),
+    siteName: getSiteName(doc),
+    description: getDescription(doc),
+    author: getAuthor(doc),
+    mediaType: getMediaType(doc) || `website`,
+    contentType,
+    images: getImages(doc, url, options.imagesPropertyType),
+    videos: getVideos(doc),
+    favicons: getFavicons(doc, url),
+  };
+
+  if (options?.onResponse && typeof options.onResponse !== `function`) {
+    throw new Error(`link-preview-js onResponse option must be a function`);
+  }
+
+  if (options?.onResponse) {
+    // send in a cloned response (to avoid mutation of original response reference)
+    const clonedResponse = structuredClone(response);
+    const urlObject = new URL(url);
+    response = options.onResponse(clonedResponse, doc, urlObject);
+  }
+
+  return response;
+}
+
+function parseUnknownResponse(
+  body: string,
+  url: string,
+  options: ILinkPreviewBaseOptions = {},
+  contentType?: string,
+) {
+  return parseTextResponse(body, url, options, contentType);
+}
+
+export function parseResponse(response: IPreFetchedResource, options?: ILinkPreviewBaseOptions) {
+  try {
+    let contentType = response.headers[`content-type`];
+    let contentTypeTokens: string[] = [];
+    let charset = null;
+
+    if (!contentType) {
+      return parseUnknownResponse(response.data, response.url, options);
+    }
+
+    if (contentType.includes(`;`)) {
+      contentTypeTokens = contentType.split(`;`);
+      contentType = contentTypeTokens[0];
+
+      for (let token of contentTypeTokens) {
+        if (token.indexOf("charset=") !== -1) {
+          charset = token.split("=")[1];
+        }
+      }
+    }
+
+    // parse response depending on content type
+    if (CONSTANTS.REGEX_CONTENT_TYPE_IMAGE.test(contentType)) {
+      return { ...parseImageResponse(response.url, contentType), charset };
+    }
+
+    if (CONSTANTS.REGEX_CONTENT_TYPE_AUDIO.test(contentType)) {
+      return { ...parseAudioResponse(response.url, contentType), charset };
+    }
+
+    if (CONSTANTS.REGEX_CONTENT_TYPE_VIDEO.test(contentType)) {
+      return { ...parseVideoResponse(response.url, contentType), charset };
+    }
+
+    if (CONSTANTS.REGEX_CONTENT_TYPE_TEXT.test(contentType)) {
+      return {
+        ...parseTextResponse(response.data, response.url, options, contentType),
+        charset,
+      };
+    }
+
+    if (CONSTANTS.REGEX_CONTENT_TYPE_APPLICATION.test(contentType)) {
+      return {
+        ...parseApplicationResponse(response.url, contentType),
+        charset,
+      };
+    }
+
+    const htmlString = response.data;
+
+    return {
+      ...parseUnknownResponse(htmlString, response.url, options),
+      charset,
+    };
+  } catch (e) {
+    throw new Error(`link-preview-js could not fetch link information ${(e as any).toString()}`);
+  }
+}
+
+/**
+ * Skip the library fetching the website for you, instead pass a response object
+ * from whatever source you get and use the internal parsing of the HTML to return
+ * the necessary information
+ * @param response Preview Response
+ * @param options IPreviewLinkOptions
+ */
+export async function getPreviewFromContent(
+  response: IPreFetchedResource,
+  options?: ILinkPreviewBaseOptions,
+) {
+  if (!response || typeof response !== `object`) {
+    throw new Error(`link-preview-js did not receive a valid response object`);
+  }
+
+  if (!response.url) {
+    throw new Error(`link-preview-js did not receive a valid response object`);
+  }
+
+  return parseResponse(response, options);
+}
